chore: rename navItem -> nav-items
